+++ src/player/src/actions/endBackgroundAudio.ts
@@ -1,0 +1,11 @@
+import { ActionForPlayback } from "../typings";
+import { endBackgroundItem } from "../media";
+
+const action = {
+  render: (action: ActionForPlayback, done: Function) => {
+    endBackgroundItem("audio", action.props.name);
+    done({});
+  },
+};
+
+export { action };